[Benchmark]Restructure the benchmark folder
This patch re-arranged the benchmark folder, which includes:

- Remove the original endurance tests.
- Classify the benchmark by different directories.
- Update the README.
- Use colcon instead of ament to build the rclcpp benchmark.

Fix #NONE
