--- benchmark/rclcpp/subscription-endurance-test.cpp
@@ -1,27 +1,0 @@
-// Copyright (c) 2018 Intel Corporation. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "rclcpp/rclcpp.hpp"
-#include "sensor_msgs/msg/joint_state.hpp"
-
-int main(int argc, char* argv[]) {
-  rclcpp::init(argc, argv);
-  auto node = rclcpp::Node::make_shared("endurance_subscription_rclcpp");
-  auto sub = node->create_subscription<sensor_msgs::msg::JointState>(
-      "endurance_topic",
-      [](sensor_msgs::msg::JointState::SharedPtr msg) { (void)msg; });
-  rclcpp::spin(node);
-
-  return 0;
-}
