--- benchmark/rclpy/service-endurance-test.py
@@ -1,33 +1,0 @@
-#!/usr/bin/env python3
-# Copyright (c) 2018 Intel Corporation. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import rclpy
-from std_srvs.srv import *
-
-def callback(request, response):
-  response.success = True
-  response.message = 'The flag has been set.'
-  return response
-
-def main():
-  rclpy.init()
-  node = rclpy.create_node('endurance_service_rclpy')
-  service = node.create_service(SetBool, 'set_flag', callback)
-
-  while rclpy.ok():
-    rclpy.spin_once(node)
-
-if __name__ == '__main__':
-  main()
